HPCC-7946 CATCH(expr,msg) won't compile
Rearrange jlib headers so that IException is not an opaque type.
