+++ system/jlib/jcrc.hpp
@@ -24,7 +24,7 @@
 
 #include "platform.h"
 #include <stdio.h>
-#include "jexpdef.hpp"
+#include "jiface.hpp"
 #include "jio.hpp"
