+++ system/jlib/jhash.ipp
@@ -21,7 +21,7 @@
 #define JHASH_IPP
 
 #include "platform.h"
-#include "jexpdef.hpp"
+#include "jiface.hpp"
 #include "jobserve.ipp"
 
 #ifdef _WIN32
