+++ system/jlib/jsorta.hpp
@@ -21,7 +21,7 @@
 #define JSORTARR_HPP
 
 
-#include "jexpdef.hpp"
+#include "jiface.hpp"
 #include "jsort.hpp"
